+++ docs/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - Fixed displaying 0.0% RPM set point after reset
 - Increased start duty cycle for velocity mode from 4% to 15%
-- Max. PWM config option
+- Max. PWM config option instead of the current limit
 - Tune PID settings menu while motor is running
+- Dynamic averaging of the RPM sense pulses depending on the set RPM
 - Improved menu acceleration
 
 ## 1.0.1